requirements.txt
selenium
pytest
webdriver-manager
pyyaml
pytest-html
pytest-xdist
python-dotenv